refactor(utils): consolidate relay registry helpers into registry.go
Move ResolvePortalRelayURLs and fetchRelayRegistry next to
LoadLocalRelayRegistry and share a single relayRegistry envelope type
so the registry domain lives in one file instead of being scattered
across network.go.
